+++ crates/objc2/src/ffi/protocol.rs
@@ -5,45 +5,32 @@ use std::os::raw::c_uint;
 #[cfg(any(doc, not(feature = "unstable-objfw")))]
 use crate::ffi::{objc_method_description, objc_property, objc_property_attribute_t};
 use crate::{
-    ffi::{OpaqueData, BOOL},
-    runtime::Sel,
+    ffi::BOOL,
+    runtime::{AnyProtocol, Sel},
 };
-
-/// Opaque type for Objective-C protocols.
-///
-/// Note that, although protocols are objects, sending messages to them is
-/// deprecated and may not work in the future.
-///
-/// The naming of this follows GNUStep; this struct does not exist in Apple's
-/// runtime, there `Protocol` is a type alias of `objc_object`.
-#[repr(C)]
-pub struct objc_protocol {
-    _priv: [u8; 0],
-    _p: OpaqueData,
-}
 
 extern_c! {
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn objc_getProtocol(name: *const c_char) -> *const objc_protocol;
+    pub fn objc_getProtocol(name: *const c_char) -> *const AnyProtocol;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     /// The returned array is deallocated with [`free`][crate::ffi::free].
-    pub fn objc_copyProtocolList(out_len: *mut c_uint) -> *mut *const objc_protocol;
+    pub fn objc_copyProtocolList(out_len: *mut c_uint) -> *mut *const AnyProtocol;
 
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn objc_allocateProtocol(name: *const c_char) -> *mut objc_protocol;
+    pub fn objc_allocateProtocol(name: *const c_char) -> *mut AnyProtocol;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn objc_registerProtocol(proto: *mut objc_protocol);
+    pub fn objc_registerProtocol(proto: *mut AnyProtocol);
 
     pub fn protocol_conformsToProtocol(
-        proto: *const objc_protocol,
-        other: *const objc_protocol,
+        proto: *const AnyProtocol,
+        other: *const AnyProtocol,
     ) -> BOOL;
-    pub fn protocol_isEqual(proto: *const objc_protocol, other: *const objc_protocol) -> BOOL;
-    pub fn protocol_getName(proto: *const objc_protocol) -> *const c_char;
+    pub fn protocol_isEqual(proto: *const AnyProtocol, other: *const AnyProtocol) -> BOOL;
+    pub fn protocol_getName(proto: *const AnyProtocol) -> *const c_char;
 
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     pub fn protocol_addMethodDescription(
-        proto: *mut objc_protocol,
+        proto: *mut AnyProtocol,
         name: Sel,
         types: *const c_char,
         is_required_method: BOOL,
@@ -51,7 +38,7 @@ extern_c! {
     );
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     pub fn protocol_addProperty(
-        proto: *mut objc_protocol,
+        proto: *mut AnyProtocol,
         name: *const c_char,
         attributes: *const objc_property_attribute_t,
         attributes_len: c_uint,
@@ -59,11 +46,11 @@ extern_c! {
         is_instance_property: BOOL,
     );
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
-    pub fn protocol_addProtocol(proto: *mut objc_protocol, addition: *const objc_protocol);
+    pub fn protocol_addProtocol(proto: *mut AnyProtocol, addition: *const AnyProtocol);
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     /// The returned array is deallocated with [`free`][crate::ffi::free].
     pub fn protocol_copyMethodDescriptionList(
-        proto: *const objc_protocol,
+        proto: *const AnyProtocol,
         is_required_method: BOOL,
         is_instance_method: BOOL,
         out_len: *mut c_uint,
@@ -71,25 +58,25 @@ extern_c! {
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     /// The returned array is deallocated with [`free`][crate::ffi::free].
     pub fn protocol_copyPropertyList(
-        proto: *const objc_protocol,
+        proto: *const AnyProtocol,
         out_len: *mut c_uint,
     ) -> *mut *const objc_property;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     /// The returned array is deallocated with [`free`][crate::ffi::free].
     pub fn protocol_copyProtocolList(
-        proto: *const objc_protocol,
+        proto: *const AnyProtocol,
         out_len: *mut c_uint,
-    ) -> *mut *const objc_protocol;
+    ) -> *mut *const AnyProtocol;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     pub fn protocol_getMethodDescription(
-        proto: *const objc_protocol,
+        proto: *const AnyProtocol,
         sel: Sel,
         is_required_method: BOOL,
         is_instance_method: BOOL,
     ) -> objc_method_description;
     #[cfg(any(doc, not(feature = "unstable-objfw")))]
     pub fn protocol_getProperty(
-        proto: *const objc_protocol,
+        proto: *const AnyProtocol,
         name: *const c_char,
         is_required_property: BOOL,
         is_instance_property: BOOL,