+++ crates/objc2/CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Merged and deprecated the following `ffi` types:
   - `ffi::objc_class` is merged into `runtime::AnyClass`.
   - `ffi::objc_object` is merged into `runtime::AnyObject`.
+  - `ffi::objc_protocol` is merged into `runtime::AnyProtocol`.
   - `ffi::IMP` is merged into `runtime::Imp`.
   - `ffi::objc_method` is merged into `runtime::Method`.
   - `ffi::objc_ivar` is merged into `runtime::Ivar`.